Cambios en las clases del paquete Modelo, implentacion de la interfaz Comparable Pruebas con ficheros, y ajuste de tipos de varibales para importacion, exportacion

+++ ProyectoFinalDam_1/src/main/java/Utils/Validadores.java
@@ -92,7 +92,7 @@ public class Validadores {
         return curso != null && !curso.trim().isEmpty();
     }
 
-    public static boolean validarCreditosEcts(int creditos_ects) {
+    public static boolean validarCreditosEcts(double creditos_ects) {
         return creditos_ects > 0;
     }
